feat(splash images and featured images):
